requirements.txt
django
djangorestframework
django-filter
environs